Added a start height for the trunk, to enable clones to be made

+++ Procedural Trees/Assets/01_Scripts/DataStructs.cs
@@ -10,7 +10,7 @@ public struct TrunkData {
     public float lengthVariance;
     public float taper;
     public int baseSplits;
-    public int segmentSplits;
+    public float segmentSplits;
     public float splitAngle;
     public float splitAngleVariance;
     public int curveResolution;
